app 파일에 있는 컴포넌트들을 파일로 분리
